techpack: audio: Reimport from LA.UM.7.1.r1-17400-sm8150.0
https://github.com/hoaysly/kernel-techpack-audio

+++ techpack/audio/dsp/q6core.c
@@ -1,5 +1,4 @@
 /* Copyright (c) 2012-2019, The Linux Foundation. All rights reserved.
- * Copyright (c) 2023 Qualcomm Innovation Center, Inc. All rights reserved.
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License version 2 and
@@ -208,7 +207,7 @@ EXPORT_SYMBOL(q6core_send_uevent);
 static int parse_fwk_version_info(uint32_t *payload, uint16_t payload_size)
 {
 	size_t ver_size;
-	uint16_t num_services;
+	int num_services;
 
 	pr_debug("%s: Payload info num services %d\n",
 		 __func__, payload[4]);
